Add mock data and return Product for given ID
